backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
google-genai==1.5.0
msal>=1.35.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
chromadb==0.6.3
fastembed==0.4.2
rank-bm25==0.2.2
pypdf==5.1.0
fastapi-azure-auth>=5.2.0
pyjwt[crypto]>=2.8.0
slowapi==0.1.9
python-json-logger>=2.0.7
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.24.0
